backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

# Angel One SmartAPI
smartapi-python==1.4.1
pyotp==2.9.0
websocket-client==1.8.0
pycryptodome==3.20.0
logzero==1.7.0

# Data processing
pandas==2.2.2
numpy==1.26.4

# Voice recognition
SpeechRecognition==3.10.4
pyaudio==0.2.14

# Notifications
python-telegram-bot==21.3
twilio==9.1.1

# Async & HTTP
httpx==0.27.0
aiofiles==23.2.1

# Schedule / background tasks
apscheduler==3.10.4
